backend/requirements.txt
fastapi
uvicorn[standard]
pandas
python-dotenv
PyMuPDF
google-generativeai
openpyxl
python-multipart